move dataclass to config.py
